pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp_capabilities_checker"
version = "0.1.0"
description = "A simple MCP client for checking server capabilities"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "MCP User"}
]
dependencies = [
    "mcp[cli]>=1.6.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
mcp-capabilities-checker = "mcp_capabilities_checker.main:main" 
